🚚rename: 인증 관련 내용 auth로 분리
