--- purithm/src/main/java/com/example/purithm/user/controller/response/LoginFailureResponseDto.java
@@ -1,2 +1,0 @@
-package com.example.purithm.user.controller.response;public class LoginFailureResponseDto {
-}
